add preheader, update readme
